Commit forgotten files on move #35
